requirements.txt
# Polybot - pinned dependencies
py-clob-client>=0.17.0,<1.0.0
typer>=0.9.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
urllib3>=2.0.0,<3.0.0
